Adding support for DSSE Rekor types (#39)
https://github.com/github/package-security/issues/954

Handling the type itself is easy. However, the type also has an envelope hash that it feels like we should be checking. But it turns out envelope canonicalization is not standardized across clients, so we aren't checking that here.

When we add in the concept of the artifact (specifically the hash) we should check that matches what's in the DSSE Rekor type.