Don't fail on Rekor entry verification for untrusted entries
Like #45, as long as the threshold of expected timestamps is met, then
verification should succeed. Otherwise, entries without trust root
material should be skipped.

One benefit of having the log key ID be used to look up the correct
trust root material is that we can still error out if the signature is invalid.

Ref: #43

Signed-off-by: Hayden Blauzvern <hblauzvern@google.com>