Add option to unify signed and log timestamps
This adds the WithObserverTimestamp option to verify that a threshold is
met using both RFC3161 signed timestamps and log integrated timestamps.

This updates the verification API for the log to only verify log
inclusion proofs or SETs, so that log and timestamp verification are not
conflated.

This also fixes a bug where the integrated time is used with an
inclusion proof. This is not safe to do, since the integrated time is
not authenticated metadata without a SignedEntryTimestamp signature.

Signed-off-by: Hayden Blauzvern <hblauzvern@google.com>