include skipped signatures in VerificationResult
Addresses #48. Instead of just ignoring the signatures the trust bundle
couldn't verify, we want to pass this information back to the caller.

To avoid API changes, in order to detect the unsinged entries, for each
signature type, we loop through all available signatures and mark the
ones the weren't verified.

Signed-off-by: Gustavo Rangel <gustavorangel91@gmail.com>