Fixes #257
When signing, if you optionally supply a trusted root we will attempt to
verify the bundle before we return it.

Previously we were using the wrong artifact digest for DSSE signing. We
could add a way to specify the artifact referred to in DSSE, but we are
already trusting the certificate identity in the user-supplied id token.

Signed-off-by: Zach Steindler <steiza@github.com>