Update BaseSignedEntity interface implementation (#333)
The SignedEntity interface has gone through a few iterations. The
BaseSignedEntity struct was not keeping up with it, which wasn't caught
because it is never used in sigstore-go. This change updates the struct
to implement the interface it was meant to implement, and adds an
assignment statement to make sure future errors in updates are caught at
compile time.

Signed-off-by: Colleen Murphy <colleenmurphy@google.com>