Verify certificate validity with only current time, bump conformance tests (#277)
* Verify certificate validity with only current time

Using a certificate's NBF will always pass the time verification. We
should be using only the current time to try to verify a certificate's
validity. This is likely to only work with long-lived certificates or
where verification happens immediately after signing.

Fixes #276

Signed-off-by: Hayden Blauzvern <hblauzvern@google.com>

* Change policy flag name, rename private fields

WithoutAnyObserverTimestampsUnsafe has been renamed to WithCurrentTime
as it is not unsafe to verify without a timestamp and can be useful for
private deployments.

Also renamed private fields in VerifierConfig to align naming with other
fields, with the require prefix.

Signed-off-by: Hayden Blauzvern <hblauzvern@google.com>

* Bump conformance

Signed-off-by: Hayden Blauzvern <hblauzvern@google.com>

* Clean up conformance

Signed-off-by: Hayden Blauzvern <hblauzvern@google.com>

---------

Signed-off-by: Hayden Blauzvern <hblauzvern@google.com>