Add support for SigningConfig
Created a simple wrapper around the protobuf SigningConfig. I mirrored
the existing code for TrustedRoot, though this is far simpler since
this is just a few strings.

I didn't add support for ClientTrustConfig since we won't distribute
this file via TUF, but this would be easy enough to add later on.

Fixes #185

Signed-off-by: Hayden Blauzvern <hblauzvern@google.com>