chore: relax go directive to permit 1.22.x
Now that sigstore/sigstore#1878 and sigstore/rekor#2323 have both landed
in tagged release we can pull those updates into sigstore-go and
similarly relax the go.mod go directive here to 1.22.0 as well.

The go directive sets the minimum version of Go required to use this
module. There's no need to set this to the latest semver patch version
of a given Go release unless the semantics of that version of Go are
required to build/test/use the module.

Ref: https://go.dev/ref/mod#go-mod-file-go

Signed-off-by: Dominic Evans <dominic.evans@uk.ibm.com>