signing config: Handle optional valid_for.end
then end time in valid_for is optional, and we care about whether the
value was included or not: Treat it as *Time instead of Time.

This change is debatable as it makes a lot of code uglier: the
alternative is to not do this and continue treating 1970-01-01 as a
magic value that means end time was not set.

Signed-off-by: Jussi Kukkonen <jkukkonen@google.com>