Select highest API version with multiple SigningConfig services
Right now, when selecting a service or set of services, the SigningConfig
will select the first match given a validity period and API version. If
the client supports two API versions concurrently (e.g. Rekor v1 and
Rekor v2) and the distributed signing config has entries for multiple API
versions (which it will for Rekor), then the client may select an older
API version depending on the sorting of the services.

This change sorts the supported API versions to select the highest API
version first.

Fixes #453

Signed-off-by: Hayden B <8418760+haydentherapper@users.noreply.github.com>