Allow public keys to sign hashedrekord
There was a restriction when signing Rekor v1 hashedrekord entries that
required them to use a certificate rather than a public key. This isn't
a restriction imposed by Rekor, and it is going to need to have this
kind of flexibility if and when this gets used for signing in cosign.
This change eliminates the restriction and then updates the examples to
show how this could be used.

Signed-off-by: Colleen Murphy <colleenmurphy@google.com>