Add MarshalJSON to SigningConfig, fix marshaling bug
This adds a MarshalJSON method to SigningConfig so that clients don't
have to construct the proto structure themselves. This also fixes a bug
in ToServiceProtobuf where we didn't handle the validity window. Test
cases have been updated as well to specify validity windows and
operators, and I've verified that these tests fail without the bug fix.

I've also added a vscode settings file to incldue the e2e tag for Go
builds.

Signed-off-by: Hayden B <8418760+haydentherapper@users.noreply.github.com>