Add MarshalJSON to SigningConfig, fix marshaling bug (#498)
This adds a MarshalJSON method to SigningConfig so that clients don't
have to construct the proto structure themselves. This also fixes a bug
in ToServiceProtobuf where we didn't handle the validity window. Test
cases have been updated as well to specify validity windows and
operators, and I've verified that these tests fail without the bug fix.

Signed-off-by: Hayden B <8418760+haydentherapper@users.noreply.github.com>