Allow no timestamps to be provided when verifying a key (#510)
This adds back an option that allows no timestamps to be provided for
verification. This will only work when verifying with a key, and will
throw an error if set when trying to verify a certificate.

This is also used when verifying while signing. If a log is used without
a timestamp authority, we either verify a certificate with the
integrated timestamp or no timestamp for a key. If no log or timestamp
is provided, we use current time for a certificate or no timestamp for a
key.

Fixes #501
Fixes #502

Signed-off-by: Hayden B <8418760+haydentherapper@users.noreply.github.com>
Co-authored-by: Hayden B <8418760+haydentherapper@users.noreply.github.com>